+++ sparkle/types/status.py
@@ -5,14 +5,14 @@ from enum import Enum
 
 class SolverStatus(str, Enum):
     """Possible return states for solver runs."""
-    SUCCESS = "SUCCESS"
+    SUCCESS = "SUCCESS"  # Positive status
+    UNKNOWN = "UNKNOWN"  # Semi positive status
+    SAT = "SAT"  # SAT specific positive status
+    UNSAT = "UNSAT"  # SAT specific positive status
+
+    # Negative status
     CRASHED = "CRASHED"
     TIMEOUT = "TIMEOUT"
     WRONG = "WRONG"
-    UNKNOWN = "UNKNOWN"
     ERROR = "ERROR"
     KILLED = "KILLED"
-
-    # SAT specific status
-    SAT = "SAT"
-    UNSAT = "UNSAT"